Really do conditional compilation
