+++ src/libraries/System.Net.Quic/src/Interop/MsQuicEnums.cs
