+++ src/libraries/System.Net.Quic/src/Interop/MsQuicNativeMethods.cs
